Directory reorganization & enhancements: - added new path folders: lib, ui. - recompiled and fixed page.tsx and Drawer.tsx
