+++ src/app/layout.tsx
@@ -1,14 +1,15 @@
 import type {Metadata} from "next";
 import {Inter} from "next/font/google";
-import "./globals.css";
+import "./ui/globals.css";
 import {AppRouterCacheProvider} from "@mui/material-nextjs/v13-appRouter";
 import {ThemeProvider} from "@mui/material/styles";
 import theme from "../../theme";
+import React from "react";
 
 const inter = Inter({subsets: ["latin"]});
 
 export const metadata: Metadata = {
-    title: "Slitherlink",
+    title: "SlitherLink",
     description: "The best game in the world!",
 };
