+++ src/app/ui/globals.css
@@ -28,8 +28,3 @@ body {
     ) rgb(var(--background-start-rgb));
 }
 
-@layer utilities {
-    .text-balance {
-        text-wrap: balance;
-    }
-}
